Reorganise repo structure
To make it much clearer which is part of the _library_ and which is part of the
_test app_. Refactor done by claude code / opus 4.6. Note: this will need
a change in GitHub pages setup to use the `test-app` directory.
