+++ vite.config.js
@@ -1,13 +1,14 @@
 import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react';
 
-// Test app config (the SPA at src/index.js used to develop the library
+// Test app config (the SPA at test-app/index.jsx used to develop the library
 // against canonical or local data).
 export default defineConfig({
+  root: 'test-app',
   base: './',
   plugins: [react()],
   build: {
-    outDir: 'build',
+    outDir: '../build',
     emptyOutDir: true,
   },
 });
